test-suite/python/pyproject.toml
[project]
name = "devcontainer-python-suite"
version = "0.1.0"
description = "Connectivity checks for devcontainer services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary==2.9.10",
    "mysql-connector-python==9.1.0",
    "redis==5.0.7",
    "kafka-python==2.2.15",
    "opensearch-py==2.7.1",
    "minio==7.2.7",
]

[tool.uv]
package = false
